move make terrain function inside datasets module init
